chore: make signature slot Slot types

+++ beacon-chain/blockchain/process_block.go
@@ -250,7 +250,7 @@ func (s *Service) onBlock(ctx context.Context, signed interfaces.SignedBeaconBlo
 		data := ethpbv2.OptimisticUpdate{
 			AttestedHeader: attestedHeader,
 			SyncAggregate:  &syncAggregateV1,
-			SignatureSlot:  uint64(prevSlot),
+			SignatureSlot:  prevSlot,
 		}
 
 		// Return the result
@@ -372,7 +372,7 @@ func (s *Service) onBlock(ctx context.Context, signed interfaces.SignedBeaconBlo
 					FinalizedHeader: finalizedHeader,
 					FinalityBranch:  finalityBranch,
 					SyncAggregate:   &syncAggregateV1,
-					SignatureSlot:   uint64(signed.Block().Slot()),
+					SignatureSlot:   signed.Block().Slot(),
 				}
 
 				// Return the result
